requirements.txt
python-telegram-bot>=20.0
gspread
google-auth
matplotlib
environs
collections
uuid
os
